requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
parso==0.8.4
python-dotenv>=1.0.0
#google-cloud-aiplatform>=1.38.0
#google-auth>=2.23.0
google-genai>=0.8.0
langchain-google-genai>=2.0.6
# Pretty console logging (optional)
rich>=13.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic >= 2.00.0
pydantic-settings==2.10.1
pydantic-core==2.27.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.24.0

# Pretty console logging
rich==13.9.4